+++ CHANGELOG.md
@@ -4,6 +4,18 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [unreleased]
 
+## [2.49.0] - 2020-01-03
+
+Features:
+- Use `styled-components` instead of `Radium` to `LinkBox` component.
+- Use `styled-components` instead of `Radium` to `ButtonImageWithText`.
+- Add `HEADER_HEIGHT` export to `Header` component.
+
+## [2.48.1] - 2019-12-20
+
+Fix:
+- Update margin on project Informations on `CrowdfundingCard`.
+
 ## [2.48.0] - 2019-12-17
 
 Feature: